ChessEngine (MyApp v0.1.0)

Documentation for ChessEngine.

board_from_pgn(pgn_str)

Makes a board from a pgn_str

/ 	 End of a file

Integer 	 Empty places

Notation [p, r, k, n, q, b, P, R, K, N, Q, B] 	 Piece

/ => End of a file
number => Empty places
notation [p, r, k, n, q, b, P, R, K, N, Q, B] => Piece

Examples

  iex> ChessEngine.board_from_pgn("")
  %ChessEngine.Board{ board: [
    :empty,:empty,:empty,:empty,:empty,:empty,:empty,:empty, 
    :empty,:empty,:empty,:empty,:empty,:empty,:empty,:empty, 
    :empty,:empty,:empty,:empty,:empty,:empty,:empty,:empty, 
    :empty,:empty,:empty,:empty,:empty,:empty,:empty,:empty,
    :empty,:empty,:empty,:empty,:empty,:empty,:empty,:empty,
    :empty,:empty,:empty,:empty,:empty,:empty,:empty,:empty,
    :empty,:empty,:empty,:empty,:empty,:empty,:empty,:empty,
    :empty,:empty,:empty,:empty,:empty,:empty,:empty,:empty
  ]}

init_standard_board()

Makes a standard chess game starting position

iex> ChessEngine.init_standard_board()